Complete translate_pubsub sample.
