+++ appengine-standard-java8/translate-pubsub/pom.xml
@@ -22,19 +22,16 @@
   <artifactId>appengine-translate-pubsub</artifactId>
 
   <parent>
-    <!-- Get parameters and allow bulk testing -->
-    <artifactId>getting-started-java</artifactId>
-    <groupId>com.example</groupId>
-    <version>1.0.0</version>
-    <relativePath>../..</relativePath>
+    <groupId>com.google.cloud.samples</groupId>
+    <artifactId>shared-configuration</artifactId>
+    <version>1.0.8</version>
   </parent>
 
   <properties>
     <maven.compiler.target>1.8</maven.compiler.target>
     <maven.compiler.source>1.8</maven.compiler.source>
     <failOnMissingWebXml>false</failOnMissingWebXml> <!-- REQUIRED -->
-    <appengine.maven.plugin>1.3.2</appengine.maven.plugin>
-    <jetty>9.4.4.v20170414</jetty>
+    <jetty>9.3.9.v20160517</jetty>
   </properties>
 
   <dependencies>
@@ -71,7 +68,7 @@
       <plugin>
         <groupId>com.google.cloud.tools</groupId>
         <artifactId>appengine-maven-plugin</artifactId>
-        <version>${appengine.maven.plugin}</version>
+        <version>1.3.2</version>
         <configuration>
           <!-- deploy configuration -->
         </configuration>
