--- appengine-standard-java8/translate-pubsub/sample-message.json
@@ -1,1 +1,0 @@
-{"message":{"data":"Hello again from AppEngine Standard!","attributes":{},"messageId":"91010751788941","publishTime":"2018-01-01T23:16:42.302Z","sourceLang":"en","targetLang":"fr"}}
