+++ appengine-standard-java8/translate-pubsub/sample_message.json
@@ -1,0 +1,1 @@
+{"message":{"data":"eyJkYXRhIjoidHJhbnNsYXRlIiwic291cmNlTGFuZyI6ImVuIiwidGFyZ2V0TGFuZyI6ImVuIn0=","attributes":{"sourceLang":"en","targetLang":"es"},"messageId":"181789827785060","publishTime":"2018-01-06T00:41:01.839Z"}}
