--- appengine-standard-java8/translate-pubsub/src/main/java/com/example/appengine/translate_pubsub/PubSubPush.java
@@ -1,55 +1,0 @@
-/**
- * Copyright 2018 Google Inc.
- *
- * <p>Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
- * except in compliance with the License. You may obtain a copy of the License at
- *
- * <p>http://www.apache.org/licenses/LICENSE-2.0
- *
- * <p>Unless required by applicable law or agreed to in writing, software distributed under the
- * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.example.appengine.translate_pubsub;
-
-import java.io.IOException;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@WebServlet(value = "/pubsub/push")
-public class PubSubPush extends PubSubHttpServlet {
-
-  @Override
-  public void doPost(HttpServletRequest req, HttpServletResponse resp)
-      throws IOException, ServletException {
-    String pubsubVerificationToken = System.getenv("PUBSUB_VERIFICATION_TOKEN");
-    // Do not process message if request token does not match pubsubVerificationToken
-    if (req.getParameter("token").compareTo(pubsubVerificationToken) != 0) {
-      resp.setStatus(HttpServletResponse.SC_BAD_REQUEST);
-      return;
-    }
-    // parse message object from "message" field in the request body json
-    // decode message data from base64
-    Message message = getMessage(req);
-    try {
-      messageRepository.save(message);
-      // 200, 201, 204, 102 status codes are interpreted as success by the Pub/Sub system
-      resp.setStatus(102);
-    } catch (Exception e) {
-      resp.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  private MessageRepository messageRepository;
-
-  PubSubPush(MessageRepository messageRepository) {
-    this.messageRepository = messageRepository;
-  }
-
-  public PubSubPush() {
-    this.messageRepository = MessageRepositoryImpl.getInstance();
-  }
-}
