+++ appengine-standard-java8/translate-pubsub/src/main/java/com/example/appengine/translatepubsub/MessageRepositoryImpl.java
@@ -1,18 +1,20 @@
-/**
+/*
  * Copyright 2018 Google Inc.
  *
- * <p>Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
- * except in compliance with the License. You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- * <p>http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- * <p>Unless required by applicable law or agreed to in writing, software distributed under the
- * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing permissions and
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
  * limitations under the License.
  */
 
-package com.example.appengine.translate_pubsub;
+package com.example.appengine.translatepubsub;
 
 import com.google.cloud.datastore.Datastore;
 import com.google.cloud.datastore.DatastoreOptions;
@@ -22,16 +24,30 @@ import com.google.cloud.datastore.KeyFactory;
 import com.google.cloud.datastore.Query;
 import com.google.cloud.datastore.QueryResults;
 import com.google.cloud.datastore.StructuredQuery;
+
 import java.util.ArrayList;
 import java.util.List;
 
-/** Storage for Message objects using Cloud Datastore. */
+/**
+ * Storage for Message objects using Cloud Datastore.
+ */
 public class MessageRepositoryImpl implements MessageRepository {
 
   private static MessageRepositoryImpl instance;
 
   private String messagesKind = "messages";
   private KeyFactory keyFactory = getDatastoreInstance().newKeyFactory().setKind(messagesKind);
+
+  private MessageRepositoryImpl() {
+  }
+
+  // retrieve a singleton instance
+  public static synchronized MessageRepositoryImpl getInstance() {
+    if (instance == null) {
+      instance = new MessageRepositoryImpl();
+    }
+    return instance;
+  }
 
   @Override
   public void save(Message message) {
@@ -52,11 +68,11 @@ public class MessageRepositoryImpl implements MessageRepository {
     }
 
     if (message.getSourceLang() != null) {
-        messageEntityBuilder = messageEntityBuilder.set("sourceLang", message.getSourceLang());
+      messageEntityBuilder = messageEntityBuilder.set("sourceLang", message.getSourceLang());
     }
 
     if (message.getTargetLang() != null) {
-        messageEntityBuilder = messageEntityBuilder.set("targetLang", message.getTargetLang());
+      messageEntityBuilder = messageEntityBuilder.set("targetLang", message.getTargetLang());
     }
     datastore.put(messageEntityBuilder.build());
   }
@@ -105,16 +121,5 @@ public class MessageRepositoryImpl implements MessageRepository {
   private Datastore getDatastoreInstance() {
     DatastoreOptions instance = DatastoreOptions.getDefaultInstance();
     return instance.getService();
-  }
-
-  private MessageRepositoryImpl() {
-  }
-
-  // retrieve a singleton instance
-  public static synchronized  MessageRepositoryImpl getInstance() {
-    if (instance == null) {
-      instance = new MessageRepositoryImpl();
-    }
-    return instance;
   }
 }
