+++ appengine-standard-java8/translate-pubsub/src/main/java/com/example/appengine/translatepubsub/PubSubHome.java
@@ -1,18 +1,20 @@
-/**
+/*
  * Copyright 2018 Google Inc.
  *
- * <p>Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
- * except in compliance with the License. You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
- * <p>http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
- * <p>Unless required by applicable law or agreed to in writing, software distributed under the
- * License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- * express or implied. See the License for the specific language governing permissions and
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
  * limitations under the License.
  */
 
-package com.example.appengine.translate_pubsub;
+package com.example.appengine.translatepubsub;
 
 import java.util.List;
 
@@ -20,6 +22,9 @@ public class PubSubHome {
 
   private static MessageRepository messageRepository = MessageRepositoryImpl.getInstance();
   private static int MAX_MESSAGES = 10;
+
+  private PubSubHome() {
+  }
 
   /**
    * Retrieve received messages in html.
@@ -48,6 +53,4 @@ public class PubSubHome {
   private static void addColumn(StringBuilder sb, String content) {
     sb.append("<td>").append(content).append("</td>");
   }
-
-  private PubSubHome() { }
 }
