+++ appengine-standard-java8/translate-pubsub/src/main/webapp/WEB-INF/appengine-web.xml
@@ -1,10 +1,9 @@
 <appengine-web-app xmlns="http://appengine.google.com/ns/1.0">
   <threadsafe>true</threadsafe>
   <runtime>java8</runtime>
-  <application>dzlier-work</application>
 
   <env-variables>
-    <env-var name="PUBSUB_TOPIC" value="pubsub-translate" />
-    <env-var name="PUBSUB_VERIFICATION_TOKEN" value="b11d0243-72c3-4da8-af2b-56ce93dc194a" />
+    <env-var name="PUBSUB_TOPIC" value="your-topic" />
+    <env-var name="PUBSUB_VERIFICATION_TOKEN" value="your-verification-token" />
   </env-variables>
 </appengine-web-app>
